refactor api clients module
